Split MapLayout
Fixes #1
